[Feat] #11 - Create SidebarViewCell
